refactored
refactored the files to better shape the project, all the files are in
the basic spaces now, next will be experimenting with the bot.
checking what can be done with the api and then after that looking into
angular - discord.js interations and changing bot settings and such.

will need to add sqlite database to store the data locally aswell